requirements.txt
alembic==1.6.5
click==8.0.1
colorama==0.4.4
dill==0.3.4
Flask==2.0.1
Flask-Caching==1.10.1
Flask-Cors==3.0.10
Flask-Migrate==3.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
gunicorn==20.1.0
h5py==2.10.0
hickle==4.0.4
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
mysqlclient==2.0.3
numpy==1.20.3
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
six==1.16.0
SQLAlchemy==1.4.18
Werkzeug==2.0.1
